+++ src/vue/modules/movements-top-bar-reit/index.vue
@@ -46,6 +46,18 @@
         </template>
 
         <!-- eslint-disable-next-line max-len -->
+        <template v-if="getModule().canRenderSubmodule(CoinpaymentsDepositModule) && asset.isCoinpayments">
+          <button
+            v-ripple
+            class="app__button-raised movements-top-bar-reit__button-raised"
+            @click="isCoinpaymentsDepositFormShown = true"
+          >
+            <i class="mdi mdi-download movements-top-bar-reit__btn-icon" />
+            {{ 'op-pages.deposit' | globalize }}
+          </button>
+        </template>
+
+        <!-- eslint-disable-next-line max-len -->
         <template v-if="getModule().canRenderSubmodule(TransferDrawerPseudoModule)">
           <button
             v-ripple
@@ -104,6 +116,28 @@
       </drawer>
     </template>
 
+    <!-- eslint-disable-next-line max-len -->
+    <template v-if="getModule().canRenderSubmodule(CoinpaymentsDepositModule) && asset.isCoinpayments">
+      <drawer :is-shown.sync="isCoinpaymentsDepositFormShown">
+        <template slot="heading">
+          {{ 'deposit-form' | globalize }}
+        </template>
+        <div>
+          <p class="movements-top-bar-reit__deposit-help-msg">
+            {{ 'deposit-form.how-to' | globalize }}
+          </p>
+          <submodule-importer
+            :submodule="getModule().getSubmodule(CoinpaymentsDepositModule)"
+            :asset="asset"
+            :balance-id="asset.balance.id"
+            :wallet="wallet"
+            :account-id="accountId"
+            :config="{horizonURL: config.horizonURL}"
+          />
+        </div>
+      </drawer>
+    </template>
+
     <drawer :is-shown.sync="isTransferDrawerShown">
       <template slot="heading">
         {{ 'transfer-form.form-heading' | globalize }}
@@ -147,6 +181,7 @@ import { TransferDrawerPseudoModule } from '@/modules-arch/pseudo-modules/transf
 import { DepositFiatModule } from '@/vue/modules/deposit-fiat/module'
 import { WithdrawalFiatModule } from '@/vue/modules/withdrawal-fiat/module'
 import { RedeemFormModule } from '@/vue/modules/redeem-form/module'
+import { CoinpaymentsDepositModule } from '@/vue/modules/coinpayments-deposit/module'
 
 const EVENTS = {
   withdrawn: 'withdrawn',
@@ -185,6 +220,7 @@ export default {
     isInitialized: false,
     isTransferDrawerShown: false,
     isFiatDepositFormShown: false,
+    isCoinpaymentsDepositFormShown: false,
     isReedemDrawerShown: false,
     isDepositDrawerShown: false,
     isFiatWithdrawalFormShown: false,
@@ -198,6 +234,7 @@ export default {
     DepositFiatModule,
     WithdrawalFiatModule,
     RedeemFormModule,
+    CoinpaymentsDepositModule,
     asset: {},
   }),
   computed: {
@@ -311,5 +348,11 @@ export default {
 .movements-top-bar-reit__filters-prefix {
   margin-right: 1.5rem;
   line-height: 1;
+}
+
+.movements-top-bar-reit__deposit-help-msg {
+  font-size: 1.2rem;
+  opacity: 0.7;
+  margin-bottom: 2rem;
 }
 </style>